refactor: extract CollectionIntegrityStorageCheck
